+++ src/clusters/raft/messages/mod.rs
@@ -1,8 +1,10 @@
+mod actor;
 mod command;
 mod event;
 mod rpc;
 mod timer;
 
+pub use actor::*;
 pub use command::*;
 pub use event::*;
 pub use rpc::*;
