+++ src/clusters/swims/messages/mod.rs
@@ -1,10 +1,13 @@
+mod actor;
 mod command;
 pub(crate) mod dissemination_buffer;
 mod packet;
 mod timer;
 
-pub use command::{MembershipEvent, SwimEvent, SwimQueryCommand};
+pub use actor::SwimQueryCommand;
+pub use command::{MembershipEvent, SwimEvent};
 
-pub(crate) use command::{SwimActorCommand, SwimCommand, SwimTimeOutCallback, SwimTimerKind};
+pub(crate) use actor::SwimActorCommand;
+pub(crate) use command::{SwimCommand, SwimTimeOutCallback, SwimTimerKind};
 pub use packet::{OutboundPacket, SwimHeader, SwimPacket};
 pub(crate) use timer::*;
